fix: guard PrivyProvider and isolate useFundWallet to prevent prerender crash
Next.js prerendering throws when PrivyProvider receives an empty appId.
Guard AppProvider so it only wraps with PrivyProvider when
NEXT_PUBLIC_PRIVY_APP_ID is set. Extract useFundWallet into a standalone
PrivyDepositButton component so the hook is never called outside the
provider. Account-menu uses a build-time PRIVY_ENABLED flag to swap
between the live button and a disabled fallback with no hydration mismatch.
